lab #8 done

+++ Lab #7/TASK_1_getRentCost.sql
@@ -8,7 +8,7 @@ CREATE FUNCTION getRentCost(memid INT, facid INT, slots INT) RETURNS DECIMAL(10,
     READS SQL DATA
     NOT DETERMINISTIC
 BEGIN
-    DECLARE income INT;
+    DECLARE income DECIMAL(10, 2);
     SET income = (
         SELECT IF(memid = 0, guestcost, membercost) * slots
         FROM facilities
